requirements.txt
fastapi
uvicorn
pymongo
pika
tenacity
python-dotenv 